move getCoverageColor to utils
