+++ __tests__/summary.test.ts
@@ -104,12 +104,3 @@ describe('should parse summary', () => {
     )
   })
 })
-
-test('should return right colors', () => {
-  expect(getCoverageColor(35)).toBe('red')
-  expect(getCoverageColor(50)).toBe('orange')
-  expect(getCoverageColor(70)).toBe('yellow')
-  expect(getCoverageColor(85)).toBe('green')
-  expect(getCoverageColor(95)).toBe('brightgreen')
-  expect(getCoverageColor(-1)).toBe('red')
-})
