+++ src/summary.ts
@@ -1,12 +1,6 @@
 import * as core from '@actions/core'
-import {
-  CoverageColor,
-  SummaryReport,
-  LineSummary,
-  Options,
-  Summary,
-} from './types.d'
-import { getContentFile } from './utils'
+import { SummaryReport, LineSummary, Options, Summary } from './types.d'
+import { getContentFile, getCoverageColor } from './utils'
 
 // parse coverage-summary.json to Sumamry object
 function parseSummary(jsonContent: string): Summary | null {
@@ -61,40 +55,6 @@ ${table}`
   }
 
   return table
-}
-
-// get coverage color
-function getCoverageColor(percentage: number): CoverageColor {
-  // https://shields.io/category/coverage
-  const rangeColors: { color: CoverageColor; range: [number, number] }[] = [
-    {
-      color: 'red',
-      range: [0, 40],
-    },
-    {
-      color: 'orange',
-      range: [40, 60],
-    },
-    {
-      color: 'yellow',
-      range: [60, 80],
-    },
-    {
-      color: 'green',
-      range: [80, 90],
-    },
-    {
-      color: 'brightgreen',
-      range: [90, 101],
-    },
-  ]
-
-  const { color } =
-    rangeColors.find(
-      ({ range: [min, max] }) => percentage >= min && percentage < max
-    ) || rangeColors[0]
-
-  return color
 }
 
 // get coverage and color from summary
