+++ src/utils.ts
@@ -1,5 +1,6 @@
 import * as core from '@actions/core'
 import { existsSync, readFileSync } from 'fs'
+import { CoverageColor } from './types'
 
 export function getPathToFile(pathToFile: string): string {
   if (!pathToFile) {
@@ -35,4 +36,38 @@ export function getContentFile(pathToFile: string): string {
 
   core.info(`File read successfully "${pathToFile}"`)
   return content
+}
+
+// get coverage color from percentage
+export function getCoverageColor(percentage: number): CoverageColor {
+  // https://shields.io/category/coverage
+  const rangeColors: { color: CoverageColor; range: [number, number] }[] = [
+    {
+      color: 'red',
+      range: [0, 40],
+    },
+    {
+      color: 'orange',
+      range: [40, 60],
+    },
+    {
+      color: 'yellow',
+      range: [60, 80],
+    },
+    {
+      color: 'green',
+      range: [80, 90],
+    },
+    {
+      color: 'brightgreen',
+      range: [90, 101],
+    },
+  ]
+
+  const { color } =
+    rangeColors.find(
+      ({ range: [min, max] }) => percentage >= min && percentage < max
+    ) || rangeColors[0]
+
+  return color
 }
